mod_panopto: Add backup and restore functionality.
When we restore the course, we only add Panopto session ID to the database
record for Panopto instance. The view functionality will create a missing
group and fix all the rights. This give us an advantage not to rely on
Panopto repo settings to perform API calls on restore. However, there are
limitations, the activity needs to be restored on the site that is linked
to the same Panopto server.